なんかこう、お何を四天王？

+++ Sources/server/internal/native/entity_data.go
@@ -2,10 +2,10 @@ package native
 
 type EntityCollision struct {
 	// Fraction is the division factor of the size of the scaling that is pre-invoked when drawing. 1 means no scaling.
-	Fraction float64 `json:"fraction"`
+	Fraction float32 `json:"fraction"`
 
 	// Radius is the radius of circle.
-	Radius float64 `json:"radius"`
+	Radius float32 `json:"radius"`
 }
 
 type EntityData[I18n any] struct {
