+++ Sources/server/internal/wave/entity_collision.go
@@ -2,6 +2,6 @@ package wave
 
 import "flooooio/internal/native"
 
-func CalculateSearchRadius(collision native.EntityCollision, size float64) float64 {
+func CalculateSearchRadius(collision native.EntityCollision, size float32) float32 {
 	return (2 * collision.Radius) * (size / collision.Fraction)
 }
